restructure page to allow server side rendering
